\name{FuzzyNumbers.Ext.2-package}
\alias{FuzzyNumbers.Ext.2-package}
\alias{FuzzyNumbers.Ext.2}
\docType{package}
\title{
\packageTitle{FuzzyNumbers.Ext.2}
}
\description{
\packageDescription{FuzzyNumbers.Ext.2}
}
\details{
The main goal of Package \code{FuzzyNumbers.Ext.2} is apply two fuzzy numbers \eqn{x} and \eqn{y} into a monotone two-variable function \eqn{f(x,y)} which is possible by using function \code{f2apply}.
}
\author{
\packageMaintainer{FuzzyNumbers.Ext.2}
}
\references{
Gagolewski, M., Caha, J., FuzzyNumbers Package: Tools to Deal with Fuzzy Numbers in R. R package version 0.4-1, 2015. https://cran.r-project.org/web/packages=FuzzyNumbers

Klir, G.J., Yuan, B., Fuzzy Sets and Fuzzy Logic: Theory and Applications, Prentice Hall PTR, New Jersey (1995).

Viertl, R., Statistical methods for fuzzy data. New York: John Wiley & Sons (2011)

Zadeh, L.A., Fuzzy sets. Information and Control 8, 338-359 (1965)

Zadeh, L.A., Probability measures of fuzzy events. Journal of Mathematical Analysis and Applications 23, 421-427 (1968)
}

\keyword{ fapply }
\keyword{ f2apply }
\keyword{ monoton function }
\keyword{ is.increasing }
\keyword{ is.decreasing }
\keyword{ is.increasing.on.x }
\keyword{ is.decreasing.on.x }
\keyword{ is.increasing.on.y }
\keyword{ is.decreasing.on.y }

\seealso{
\code{FuzzyNumbers}
}
